handle null / missing columns properly
Pull request #7 has an issue where rows that contain missing / null values will have their columns out of alignment with the data in other columns, and with the header/footer.

This resolves that issue by adding a string as a placeholder for null or missing columns. The value of this string can be changed via the $_null configuration option, and defaults to 'NULL'.

I think this code should also resolve issue #6.

Signed-off-by: Joshua Paling <joshua.paling@gmail.com>